[PASSBE-143] Test Core Data models

+++ Client/Models/Share.swift
@@ -28,13 +28,13 @@ public struct Share: Codable {
     public let vaultID: String
 
     /// Type of share. 1 for vault, 2 for label and 3 for item
-    public let targetType: Int
+    public let targetType: Int16
 
     /// ID of the top shared object
     public let targetID: String
 
     /// Permissions for this share
-    public let permission: Int
+    public let permission: Int16
 
     /// Base64 encoded signature of the vault signing key fingerprint
     public let acceptanceSignature: String
@@ -70,19 +70,19 @@ public struct Share: Codable {
     public let contentSignatureEmail: String
 
     /// Version of the content's format
-    public let contentFormatVersion: Int
+    public let contentFormatVersion: Int16
 
     /// Expiration time for this share
-    public let expireTime: Double?
+    public let expireTime: Int64?
 
     /// Time of creation of this share
-    public let createTime: Double
+    public let createTime: Int64
 
     public init(shareID: String,
                 vaultID: String,
-                targetType: Int,
+                targetType: Int16,
                 targetID: String,
-                permission: Int,
+                permission: Int16,
                 acceptanceSignature: String,
                 inviterEmail: String,
                 inviterAcceptanceSignature: String,
@@ -93,9 +93,9 @@ public struct Share: Codable {
                 contentEncryptedAddressSignature: String,
                 contentEncryptedVaultSignature: String,
                 contentSignatureEmail: String,
-                contentFormatVersion: Int,
-                expireTime: Double?,
-                createTime: Double) {
+                contentFormatVersion: Int16,
+                expireTime: Int64?,
+                createTime: Int64) {
         self.shareID = shareID
         self.vaultID = vaultID
         self.targetType = targetType
